Trait xcm_executor::traits::CheckSuspension

pub trait CheckSuspension {
    fn is_suspended<Call>(
        origin: &MultiLocation,
        instructions: &mut [Instruction<Call>],
        max_weight: Weight,
        weight_credit: &mut Weight
    ) -> bool;
}

Trait to determine whether the execution engine is suspended from executing a given XCM.

The trait method is given the same parameters as ShouldExecute::should_execute, so that the implementer will have all the context necessary to determine whether or not to suspend the XCM executor.

Can be chained together in tuples to have multiple rounds of checks. If all of the tuple elements returns false, then execution is not suspended. Otherwise, execution is suspended if any of the tuple elements returns true.

Required Methods

fn is_suspended<Call>(
origin: &MultiLocation,
instructions: &mut [Instruction<Call>],
max_weight: Weight,
weight_credit: &mut Weight
) -> bool