Trait xcm_executor::traits::FeeManager

pub trait FeeManager {
    fn is_waived(origin: Option<&MultiLocation>, r: FeeReason) -> bool;
    fn handle_fee(fee: MultiAssets);
}

Handle stuff to do with taking fees in certain XCM instructions.

Required Methods

fn is_waived(origin: Option<&MultiLocation>, r: FeeReason) -> bool

Determine if a fee which would normally payable should be waived.

fn handle_fee(fee: MultiAssets)

Do something with the fee which has been paid. Doing nothing here silently burns the fees.