Trait xcm_executor::traits::WeightBounds

pub trait WeightBounds<RuntimeCall> {
    fn weight(message: &mut Xcm<RuntimeCall>) -> Result<Weight, ()>;
    fn instr_weight(
        instruction: &Instruction<RuntimeCall>
    ) -> Result<Weight, ()>;
}

Determine the weight of an XCM message.

Required Methods

fn weight(message: &mut Xcm<RuntimeCall>) -> Result<Weight, ()>

Return the maximum amount of weight that an attempted execution of this message could consume.

fn instr_weight(instruction: &Instruction<RuntimeCall>) -> Result<Weight, ()>

Return the maximum amount of weight that an attempted execution of this instruction could consume.