Module xcm_executor::traits

Various traits used in configuring the executor.

Modules

prelude

Structs

Identity

A structure that performs identity conversion.

JustTry

A structure that performs standard conversion using the standard Rust conversion traits.

Properties

Properyies of an XCM message and its imminent execution.

WithOriginFilter

Enums

Error

Errors associated with MatchesFungibles operation.

FeeReason

Context under which a fee is paid.

LockError

QueryResponseStatus

The possible state of an XCM query response.

Traits

AssetExchange

A service for exchanging assets.

AssetLock

Define a handler for notification of an asset being locked and for the unlock instruction.

CallDispatcher

Defines how a call is dispatched with given origin. Allows to customize call dispatch, such as adapting the origin based on the call or modifying the call.

CheckSuspension

Trait to determine whether the execution engine is suspended from executing a given XCM.

ClaimAssets

Define any handlers for the AssetClaim instruction.

ConvertLocation

Means of converting a location into an account identifier.

ConvertOrigin

A converter trait for origin types.

DropAssets

Define a handler for when some non-empty Assets value should be dropped.

Enact

ExportXcm

Utility for delivering a message to a system under a different (non-local) consensus with a spoofed origin. This essentially defines the behaviour of the ExportMessage XCM instruction.

FeeManager

Handle stuff to do with taking fees in certain XCM instructions.

FilterAssetLocationDeprecated

Filters assets/location pairs.

MatchesFungible

MatchesFungibles

MatchesNonFungible

MatchesNonFungibles

OnResponse

Define what needs to be done upon receiving a query response.

QueryHandler

Provides methods to expect responses from XCMs and query their status.

ShouldExecute

Trait to determine whether the execution engine should actually execute a given XCM.

TransactAsset

Facility for asset transacting.

VersionChangeNotifier

Trait for a type which handles notifying a destination of XCM version changes.

WeightBounds

Determine the weight of an XCM message.

WeightTrader

Charge for weight in order to execute XCM.

Functions

export_xcm

Convenience function for using a SendXcm implementation. Just interprets the dest and wraps both in Some before passing them as as mutable references into T::send_xcm.

validate_export

Convenience function for using a SendXcm implementation. Just interprets the dest and wraps both in Some before passing them as as mutable references into T::send_xcm.