Trait xcm_executor::traits::VersionChangeNotifier

pub trait VersionChangeNotifier {
    fn start(
        location: &MultiLocation,
        query_id: QueryId,
        max_weight: u64
    ) -> XcmResult;
    fn stop(location: &MultiLocation) -> XcmResult;
    fn is_subscribed(location: &MultiLocation) -> bool;
}

Trait for a type which handles notifying a destination of XCM version changes.

Required Methods

fn start(
location: &MultiLocation,
query_id: QueryId,
max_weight: u64
) -> XcmResult

Start notifying location should the XCM version of this chain change.

When it does, this type should ensure a QueryResponse message is sent with the given query_id & max_weight and with a response of Repsonse::Version. This should happen until/unless stop is called with the correct query_id.

If the location has an ongoing notification and when this function is called, then an error should be returned.

fn stop(location: &MultiLocation) -> XcmResult

Stop notifying location should the XCM change. Returns an error if there is no existing notification set up.

fn is_subscribed(location: &MultiLocation) -> bool

Return true if a location is subscribed to XCM version changes.